revert rename examples to test, move tests from init to test_zeros
