+++ scipy/optimize/cython_optimize/tests/zeros_examples.pxd
